CrowdCounting-P2PNet/requirements.txt
torch
torchvision
tensorboardX
easydict
pandas
matplotlib
Pillow